Appendix B - Source code/modules/LinuxAudio/requirements.txt
azure-iot-device~=2.7.0
PyAudio~=0.2.11
numpy~=1.21.5